requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=1.7.0
prophet>=1.1.0
joblib>=1.3.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
requests>=2.31.0
